feat: 完善25E调度、融合控制与视觉通信

+++ 25E Code/MSPM0G3507/TemplateCar/Driver/wit_imu.h
@@ -25,12 +25,12 @@ typedef struct {
     float roll;
     float yaw;
     float temperature;
-    int16_t ax;
-    int16_t ay;
-    int16_t az;
-    int16_t gx;
-    int16_t gy;
-    int16_t gz;
+    float ax;
+    float ay;
+    float az;
+    float gx;
+    float gy;
+    float gz;
     int16_t version;
 } WIT_Data_t;
